+++ packages/@justweb3/xmtp-plugin/src/lib/components/JustWeb3ButtonRight/index.tsx
@@ -3,7 +3,8 @@ import { NotificationBadge } from '@justweb3/ui';
 import { ChatIcon } from '../../icons/ChatIcon';
 import { useJustWeb3XMTP } from '../../providers/JustWeb3XMTPProvider';
 import { useMemo } from 'react';
-import { useClient } from '@xmtp/react-sdk';
+import { useEthersSigner, useXMTPClient } from '../../hooks';
+import { useXMTPContext } from '../../hooks/useXMTPContext';
 
 export interface ChatMenuButtonProps {
   handleOpen: (open: boolean) => void;
@@ -14,16 +15,18 @@ export const JustWeb3ButtonRight: React.FC<ChatMenuButtonProps> = ({
   handleOpen,
   env,
 }) => {
-  const { conversationsInfo, initializeXmtp } = useJustWeb3XMTP();
+  const { conversationsInfo } = useJustWeb3XMTP();
+  const { initializeXmtp } = useXMTPClient();
+  const { client } = useXMTPContext();
+  const signer = useEthersSigner();
   const totalUnreadCount = useMemo(() => {
     return conversationsInfo
       .filter((conversation) => conversation.consent === 'allowed')
       .reduce((acc, curr) => acc + curr.unreadCount, 0);
   }, [conversationsInfo]);
-  const { client } = useClient();
   const handleChat = async () => {
-    if (!client) {
-      initializeXmtp().then(() => {
+    if (!client && !!signer) {
+      initializeXmtp({ signer }).then(() => {
         handleOpen(true);
       });
     } else {