+++ packages/@justweb3/xmtp-plugin/src/lib/utils/groupMessageByDate/index.ts
@@ -1,37 +1,48 @@
-import { MessageWithReaction } from '../filterReactionsMessages'
+import { MessageWithReaction } from '../filterReactionsMessages';
 
 interface GroupedMessages {
-    [date: string]: MessageWithReaction[]
+  [date: string]: MessageWithReaction[];
 }
 
 const getFormattedDate = (date: Date): string => {
-    const options: Intl.DateTimeFormatOptions = {
-        weekday: 'long',
-        day: '2-digit',
-        month: 'short',
-    }
-    const formattedDate = date
-        .toLocaleDateString('en-US', options)
-        .toUpperCase()
-    const parts = formattedDate.split(' ')
-    return `${parts[0]} ${parts[1]} ${parts[2]}`
-}
+  const options: Intl.DateTimeFormatOptions = {
+    weekday: 'long',
+    day: '2-digit',
+    month: 'short',
+  };
+  const formattedDate = date.toLocaleDateString('en-US', options).toUpperCase();
+  const parts = formattedDate.split(' ');
+  return `${parts[0]} ${parts[1]} ${parts[2]}`;
+};
+
+const convertNanosToDate = (sentAtNs: string | number | bigint): Date => {
+  if (typeof sentAtNs === 'bigint') {
+    return new Date(Number(sentAtNs / BigInt(1000000)));
+  }
+
+  if (typeof sentAtNs === 'string' && sentAtNs.endsWith('n')) {
+    const bigIntValue = BigInt(sentAtNs.slice(0, -1));
+    return new Date(Number(bigIntValue / BigInt(1000000)));
+  }
+
+  return new Date(Number(sentAtNs) / 1000000);
+};
 
 export const groupMessagesByDate = (
-    messages: MessageWithReaction[],
+  messages: MessageWithReaction[]
 ): GroupedMessages => {
-    const groupedMessages = messages.reduce<GroupedMessages>((acc, message) => {
-        const date = new Date(message.sentAt)
-        const formattedDate = getFormattedDate(date)
+  const groupedMessages = messages.reduce<GroupedMessages>((acc, message) => {
+    const date = convertNanosToDate(message.sentAtNs);
+    const formattedDate = getFormattedDate(date);
 
-        if (!acc[formattedDate]) {
-            acc[formattedDate] = []
-        }
+    if (!acc[formattedDate]) {
+      acc[formattedDate] = [];
+    }
 
-        acc[formattedDate].push(message)
+    acc[formattedDate].push(message);
 
-        return acc
-    }, {})
+    return acc;
+  }, {});
 
-    return groupedMessages
-}
+  return groupedMessages;
+};