+++ packages/@justweb3/efp-plugin/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.82 (2025-04-23)
+
+This was a version bump only for @justweb3/efp-plugin to align it with other projects, there were no code changes.
+
+## 0.1.81 (2025-04-23)
+
+This was a version bump only for @justweb3/efp-plugin to align it with other projects, there were no code changes.
+
 ## 0.1.80 (2025-02-13)
 
 This was a version bump only for @justweb3/efp-plugin to align it with other projects, there were no code changes.