+++ packages/siwens/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.23 (2025-04-23)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
+## 0.1.22 (2025-04-23)
+
+This was a version bump only for siwens to align it with other projects, there were no code changes.
+
 ## 0.1.21 (2025-02-13)
 
 This was a version bump only for siwens to align it with other projects, there were no code changes.